Move common tools to separate file
