+++ tests/components/tradfri/test_cover.py
@@ -8,11 +8,7 @@ from pytradfri.device import Device
 from pytradfri.device.blind import Blind
 from pytradfri.device.blind_control import BlindControl
 
-from homeassistant.components import tradfri
-
-from . import GATEWAY_ID
-
-from tests.common import MockConfigEntry
+from .common import setup_integration
 
 SET_POSITION_TEST_CASES = [
     # Open cover
@@ -40,24 +36,6 @@ def setup(request):
         p_2.stop()
 
     request.addfinalizer(teardown)
-
-
-async def setup_integration(hass):
-    """Load the Tradfri platform with a mock gateway."""
-    entry = MockConfigEntry(
-        domain=tradfri.DOMAIN,
-        data={
-            "host": "mock-host",
-            "identity": "mock-identity",
-            "key": "mock-key",
-            "import_groups": True,
-            "gateway_id": GATEWAY_ID,
-        },
-    )
-
-    entry.add_to_hass(hass)
-    await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
 
 
 def mock_cover(test_features=None, test_state=None, device_number=0):
