+++ tests/components/tradfri/test_light.py
@@ -8,11 +8,7 @@ from pytradfri.device import Device
 from pytradfri.device.light import Light
 from pytradfri.device.light_control import LightControl
 
-from homeassistant.components import tradfri
-
-from . import GATEWAY_ID
-
-from tests.common import MockConfigEntry
+from .common import setup_integration
 
 DEFAULT_TEST_FEATURES = {
     "can_set_dimmer": False,
@@ -98,24 +94,6 @@ def setup(request):
 async def generate_psk(self, code):
     """Mock psk."""
     return "mock"
-
-
-async def setup_integration(hass):
-    """Load the Tradfri platform with a mock gateway."""
-    entry = MockConfigEntry(
-        domain=tradfri.DOMAIN,
-        data={
-            "host": "mock-host",
-            "identity": "mock-identity",
-            "key": "mock-key",
-            "import_groups": True,
-            "gateway_id": GATEWAY_ID,
-        },
-    )
-
-    entry.add_to_hass(hass)
-    await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
 
 
 def mock_light(test_features=None, test_state=None, light_number=0):
